requirements.txt
pyside6
psycopg2-binary
python-dotenv
pandas
